+++ fj-daogen-sample/src/test/java/org/fugerit/java/daogen/sample/junit4test/model/UploadJunit4ModelTest.java
@@ -63,6 +63,7 @@ public class UploadJunit4ModelTest {
 		finder1.setId( current.getId() );
 		Assert.assertEquals( current.getId(), finder1.getId() );
 		Assert.assertNotNull( UploadFinder.newInstance( current.getId() ) );
+		Assert.assertNotNull( UploadFinder.newInstance( current ) );
 		Assert.assertNotNull( finder1 );
 	}